requirements.txt
pydantic-settings
requests
typer[all]
click==8.1.7
pyyaml
